+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import './globals.css'
 import AlchemyProvider from '@/components/AlchemyProvider'
-import { ThemeProvider } from '@/contexts/ThemeContext'
+import { Provider } from '@/components/ui/provider'
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -25,13 +25,13 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <html lang='en'>
+    <html lang='en' suppressHydrationWarning>
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        <ThemeProvider>
+        <Provider>
           <AlchemyProvider>{children}</AlchemyProvider>
-        </ThemeProvider>
+        </Provider>
       </body>
     </html>
   )